+++ package.json
@@ -207,19 +207,19 @@
       "@ember/-internals/routing/index.js": "ember-source/@ember/-internals/routing/index.js",
       "@ember/-internals/runtime/index.js": "ember-source/@ember/-internals/runtime/index.js",
       "@ember/-internals/runtime/lib/ext/rsvp.js": "ember-source/@ember/-internals/runtime/lib/ext/rsvp.js",
-      "@ember/-internals/runtime/lib/mixins/-proxy-internal.js": "ember-source/@ember/-internals/runtime/lib/mixins/-proxy-internal.js",
-      "@ember/-internals/runtime/lib/mixins/action_handler-internal.js": "ember-source/@ember/-internals/runtime/lib/mixins/action_handler-internal.js",
+      "@ember/-internals/runtime/lib/mixins/-proxy.js": "ember-source/@ember/-internals/runtime/lib/mixins/-proxy.js",
+      "@ember/-internals/runtime/lib/mixins/action_handler.js": "ember-source/@ember/-internals/runtime/lib/mixins/action_handler.js",
       "@ember/-internals/runtime/lib/mixins/comparable.js": "ember-source/@ember/-internals/runtime/lib/mixins/comparable.js",
-      "@ember/-internals/runtime/lib/mixins/container_proxy-internal.js": "ember-source/@ember/-internals/runtime/lib/mixins/container_proxy-internal.js",
-      "@ember/-internals/runtime/lib/mixins/registry_proxy-internal.js": "ember-source/@ember/-internals/runtime/lib/mixins/registry_proxy-internal.js",
-      "@ember/-internals/runtime/lib/mixins/target_action_support-internal.js": "ember-source/@ember/-internals/runtime/lib/mixins/target_action_support-internal.js",
+      "@ember/-internals/runtime/lib/mixins/container_proxy.js": "ember-source/@ember/-internals/runtime/lib/mixins/container_proxy.js",
+      "@ember/-internals/runtime/lib/mixins/registry_proxy.js": "ember-source/@ember/-internals/runtime/lib/mixins/registry_proxy.js",
+      "@ember/-internals/runtime/lib/mixins/target_action_support.js": "ember-source/@ember/-internals/runtime/lib/mixins/target_action_support.js",
       "@ember/-internals/string/index.js": "ember-source/@ember/-internals/string/index.js",
       "@ember/-internals/utility-types/index.js": "ember-source/@ember/-internals/utility-types/index.js",
       "@ember/-internals/utils/index.js": "ember-source/@ember/-internals/utils/index.js",
       "@ember/-internals/views/index.js": "ember-source/@ember/-internals/views/index.js",
       "@ember/-internals/views/lib/compat/attrs.js": "ember-source/@ember/-internals/views/lib/compat/attrs.js",
       "@ember/-internals/views/lib/compat/fallback-view-registry.js": "ember-source/@ember/-internals/views/lib/compat/fallback-view-registry.js",
-      "@ember/-internals/views/lib/mixins/action_support-internal.js": "ember-source/@ember/-internals/views/lib/mixins/action_support-internal.js",
+      "@ember/-internals/views/lib/mixins/action_support.js": "ember-source/@ember/-internals/views/lib/mixins/action_support.js",
       "@ember/-internals/views/lib/system/event_dispatcher.js": "ember-source/@ember/-internals/views/lib/system/event_dispatcher.js",
       "@ember/-internals/views/lib/system/utils.js": "ember-source/@ember/-internals/views/lib/system/utils.js",
       "@ember/-internals/views/lib/views/core-view-utils.js": "ember-source/@ember/-internals/views/lib/views/core-view-utils.js",
