+++ packages/@ember/-internals/deprecations/index.ts
@@ -161,6 +161,83 @@ export const DEPRECATIONS = {
     until: '8.0.0',
     url: 'https://deprecations.emberjs.com/id/deprecate-promise-proxy-mixin',
   }),
+  DEPRECATE_OBSERVABLE_MIXIN: deprecation({
+    id: 'deprecate-observable-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-observable-mixin',
+  }),
+  DEPRECATE_ENUMERABLE_MIXIN: deprecation({
+    id: 'deprecate-enumerable-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-enumerable-mixin',
+  }),
+  DEPRECATE_MUTABLE_ENUMERABLE_MIXIN: deprecation({
+    id: 'deprecate-mutable-enumerable-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-mutable-enumerable-mixin',
+  }),
+  DEPRECATE_EMBER_ARRAY_MIXIN: deprecation({
+    id: 'deprecate-ember-array-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-ember-array-mixin',
+  }),
+  DEPRECATE_MUTABLE_ARRAY_MIXIN: deprecation({
+    id: 'deprecate-mutable-array-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-mutable-array-mixin',
+  }),
+  DEPRECATE_NATIVE_ARRAY_MIXIN: deprecation({
+    id: 'deprecate-native-array-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-native-array-mixin',
+  }),
+  DEPRECATE_CONTROLLER_MIXIN: deprecation({
+    id: 'deprecate-controller-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-controller-mixin',
+  }),
+  DEPRECATE_ACTION_HANDLER_MIXIN: deprecation({
+    id: 'deprecate-action-handler-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-action-handler-mixin',
+  }),
+  DEPRECATE_PROXY_MIXIN: deprecation({
+    id: 'deprecate-proxy-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-proxy-mixin',
+  }),
+  DEPRECATE_CONTAINER_PROXY_MIXIN: deprecation({
+    id: 'deprecate-container-proxy-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-container-proxy-mixin',
+  }),
+  DEPRECATE_REGISTRY_PROXY_MIXIN: deprecation({
+    id: 'deprecate-registry-proxy-mixin',
+    for: 'ember-source',
+    since: { available: '7.4.0' },
+    until: '8.0.0',
+    url: 'https://deprecations.emberjs.com/id/deprecate-registry-proxy-mixin',
+  }),
 };
 
 export function deprecateUntil(message: string, deprecation: DeprecationObject) {
