+++ packages/@ember/-internals/runtime/lib/mixins/comparable.ts
@@ -1,5 +1,4 @@
-import Mixin from '@ember/object/mixin';
-import { INTERNAL_MIXIN_CREATE } from '@ember/-internals/utils/lib/internal-mixin-create';
+import { InternalMixin } from '@ember/object/mixin-internal';
 import { deprecateUntil, DEPRECATIONS } from '@ember/-internals/deprecations';
 
 /**
@@ -20,7 +19,7 @@ import { deprecateUntil, DEPRECATIONS } from '@ember/-internals/deprecations';
 interface Comparable {
   compare: ((a: unknown, b: unknown) => -1 | 0 | 1) | null;
 }
-const Comparable = Mixin[INTERNAL_MIXIN_CREATE]({
+const Comparable = InternalMixin.create({
   /**
     __Required.__ You must implement this method to apply this mixin.
