+++ packages/@ember/-internals/runtime/tests/mixins/array_test.js
@@ -7,7 +7,8 @@ import {
   arrayContentWillChange,
 } from '@ember/-internals/metal';
 import EmberObject, { get, set, computed, observer as emberObserver } from '@ember/object';
-import EmberArray, { A as emberA } from '@ember/array';
+import { A as emberA } from '@ember/array';
+import { InternalEmberArray as EmberArray } from '@ember/array/index-internal';
 import { moduleFor, AbstractTestCase, runLoopSettled } from 'internal-test-helpers';
 
 /*
