+++ packages/@ember/-internals/views/index.ts
@@ -18,6 +18,6 @@ export {
 export { default as EventDispatcher } from './lib/system/event_dispatcher';
 export { default as CoreView } from './lib/views/core_view';
 export { sendCoreViewEvent, hasCoreViewListener } from './lib/views/core-view-utils';
-export { default as ActionSupport } from './lib/mixins/action_support-internal';
+export { default as ActionSupport } from './lib/mixins/action_support';
 export { MUTABLE_CELL } from './lib/compat/attrs';
 export { default as ViewStates } from './lib/views/states';
