+++ packages/@ember/engine/instance.ts
@@ -8,8 +8,10 @@ import { assert } from '@ember/debug';
 import { default as Registry, privatize as P } from '@ember/-internals/container/lib/registry';
 import { guidFor } from '@ember/-internals/utils/lib/guid';
 import { ENGINE_PARENT, getEngineParent, setEngineParent } from './parent';
-import ContainerProxyMixin from '@ember/-internals/runtime/lib/mixins/container_proxy-internal';
-import RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy-internal';
+import type ContainerProxyMixin from '@ember/-internals/runtime/lib/mixins/container_proxy';
+import type RegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy';
+import InternalContainerProxyMixin from '@ember/-internals/runtime/lib/mixins/container_proxy-internal';
+import InternalRegistryProxyMixin from '@ember/-internals/runtime/lib/mixins/registry_proxy-internal';
 import type { InternalOwner } from '@ember/-internals/owner';
 import type Owner from '@ember/-internals/owner';
 import { type FullName, isFactory } from '@ember/-internals/owner';
@@ -53,7 +55,10 @@ export interface EngineInstanceOptions {
 // type checking, we have broken part of our public API contract. Medium-term,
 // the goal here is to `EngineInstance` simple be `Owner`.
 interface EngineInstance extends RegistryProxyMixin, ContainerProxyMixin, InternalOwner, Owner {}
-class EngineInstance extends EmberObject.extend(RegistryProxyMixin, ContainerProxyMixin) {
+class EngineInstance extends EmberObject.extend(
+  InternalRegistryProxyMixin,
+  InternalContainerProxyMixin
+) {
   /**
    @private
    @method setupRegistry
