+++ packages/@ember/enumerable/mutable.ts
@@ -1,6 +1,6 @@
-import Enumerable from '@ember/enumerable';
-import Mixin from '@ember/object/mixin';
-import { INTERNAL_MIXIN_CREATE } from '@ember/-internals/utils/lib/internal-mixin-create';
+import InternalMutableEnumerable from '@ember/enumerable/mutable-internal';
+import { DeprecatedMixin } from '@ember/object/mixin-internal';
+import { deprecateUntil, DEPRECATIONS } from '@ember/-internals/deprecations';
 
 /**
 @module ember
@@ -15,9 +15,18 @@ import { INTERNAL_MIXIN_CREATE } from '@ember/-internals/utils/lib/internal-mixi
   @namespace Ember
   @uses Enumerable
   @private
+  @deprecated Use native arrays and array methods instead.
 */
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-interface MutableEnumerable extends Enumerable {}
-const MutableEnumerable = Mixin[INTERNAL_MIXIN_CREATE](Enumerable);
+interface MutableEnumerable extends InternalMutableEnumerable {}
+const MutableEnumerable = DeprecatedMixin.create(InternalMutableEnumerable, {
+  init() {
+    this._super(...arguments);
+    deprecateUntil(
+      'The `MutableEnumerable` mixin is deprecated. Use native arrays and array methods instead.',
+      DEPRECATIONS.DEPRECATE_MUTABLE_ENUMERABLE_MIXIN
+    );
+  },
+});
 
 export default MutableEnumerable;
