+++ packages/ember/tests/service_injection_test.js
@@ -1,7 +1,7 @@
 import { getOwner } from '@ember/-internals/owner';
 import Controller from '@ember/controller';
 import Service, { service } from '@ember/service';
-import { _ProxyMixin } from '@ember/-internals/runtime';
+import _ProxyMixin from '@ember/-internals/runtime/lib/mixins/-proxy-internal';
 import { moduleFor, ApplicationTestCase } from 'internal-test-helpers';
 import { computed } from '@ember/object';
 import { precompileTemplate } from '@ember/template-compilation';
